(update) rename utiltiy class name fron NullSafe to DefensiveCopies
