+++ spring-lens-core/src/main/java/com/sdlcpro/springlens/model/bean/definition/BeanDefinitionSummary.java
@@ -2,12 +2,12 @@ package com.sdlcpro.springlens.model.bean.definition;
 
 import com.sdlcpro.springlens.model.bean.BeanRole;
 import com.sdlcpro.springlens.model.bean.LoadingMode;
-import com.sdlcpro.springlens.util.NullSafe;
+import com.sdlcpro.springlens.util.DefensiveCopies;
 import com.sdlcpro.springlens.util.Preconditions;
 
 import java.util.Map;
 
-import static com.sdlcpro.springlens.util.NullSafe.emptyIfNull;
+import static com.sdlcpro.springlens.util.DefensiveCopies.emptyIfNull;
 
 /**
  * Immutable summary of bean definition metrics across Spring application
@@ -32,10 +32,10 @@ public record BeanDefinitionSummary(
 ) {
 
     public BeanDefinitionSummary {
-        contextDistribution = NullSafe.emptyIfNull(contextDistribution);
-        scopeDistribution = NullSafe.emptyIfNull(scopeDistribution);
-        roleDistribution = NullSafe.emptyIfNull(roleDistribution);
-        loadingModeDistribution = NullSafe.emptyIfNull(loadingModeDistribution);
+        contextDistribution = DefensiveCopies.emptyIfNull(contextDistribution);
+        scopeDistribution = DefensiveCopies.emptyIfNull(scopeDistribution);
+        roleDistribution = DefensiveCopies.emptyIfNull(roleDistribution);
+        loadingModeDistribution = DefensiveCopies.emptyIfNull(loadingModeDistribution);
 
         Preconditions.isTrue(
                 totalBeanDefinitions >= 0,
